Moved common code from src to commo folder
