+++ .github/workflows/upload-testing-image-to-ecr.yaml
@@ -6,6 +6,8 @@ on:
       - main
     paths:
       - tests/**
+      - common/**
+
   workflow_dispatch:
     branches:
       - main
