+++ common/sharedServices/s3/deleteS3Object.test.ts
@@ -9,7 +9,7 @@ import 'aws-sdk-client-mock-jest'
 import {
   TEST_TEMPORARY_BUCKET_NAME,
   TEST_S3_OBJECT_KEY
-} from '../../utils/tests/testConstants'
+} from '../../../common/utils/tests/testConstants'
 
 const s3Mock = mockClient(S3Client)
