+++ common/sharedServices/s3/s3FileExists.test.ts
@@ -8,7 +8,7 @@ import 'aws-sdk-client-mock-jest'
 import {
   TEST_PERMANENT_BUCKET_NAME,
   TEST_S3_OBJECT_KEY
-} from '../../utils/tests/testConstants'
+} from '../../../common/utils/tests/testConstants'
 import { s3FileExists } from './s3FileExists'
 const s3Mock = mockClient(S3Client)
 describe('s3FileExists', () => {
