+++ common/types/s3ObjectDetails.ts
