+++ common/utils/awsSdkClients.ts
