+++ common/utils/helpers/objectToBase64.ts
