+++ jest.config.ts
@@ -4,8 +4,8 @@ const config: Config.InitialOptions = {
   coveragePathIgnorePatterns: ['/node_modules/', '/dist/'],
   preset: 'ts-jest',
   verbose: true,
-  setupFiles: ['<rootDir>/src/utils/tests/setup/testEnvVars.ts'],
-  testMatch: ['**/*.test.ts']
+  setupFiles: ['<rootDir>/common/utils/tests/setup/testEnvVars.ts'],
+  testMatch: ['**/*.test.ts', '<rootDir>/common/**/*.test.ts']
 }
 
 export default config
