+++ src/lambdas/auditMessageFirehoseReingest/handler.test.ts
@@ -1,7 +1,7 @@
 import { SQSEvent } from 'aws-lambda'
 import { when } from 'jest-when'
-import { mockLambdaContext } from '../../utils/tests/mockLambdaContext'
-import { testS3TestEvent } from '../../utils/tests/testEvents/testS3SqsEvent'
+import { mockLambdaContext } from '../../../common/utils/tests/mockLambdaContext'
+import { testS3TestEvent } from '../../../common/utils/tests/testEvents/testS3SqsEvent'
 import { deleteOrUpdateS3Objects } from './deleteOrUpdateS3Objects'
 import { getAuditEvents } from './getAuditEvents'
 import { handler } from './handler'
